coutmerge_railway_ready/requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
aiofiles
pillow
PyPDF2
reportlab
